Code refactoring
EMP-1554
