+++ packages/x-components/src/x-modules/queries-preview/store/types.ts
@@ -5,7 +5,6 @@ import { XStoreModule } from '../../../store/store.types';
 import { RequestStatus, StatusState } from '../../../store/utils/status-store.utils';
 import { QueriesPreviewConfig } from '../config.types';
 import { QueryPreviewInfo } from '../../../x-installer/index';
-import { InternalSearchRequest } from '../../search/index';
 
 /**
  * QueriesPreview store state.
@@ -54,7 +53,7 @@ export interface QueriesPreviewGetters {
    * The adapter request object for retrieving the results, or null if there is no
    * valid data to create a request.
    */
-  request: InternalSearchRequest | null;
+  selectedQueryPreview: QueryPreviewInfo | null;
   initialParams: Dictionary<unknown>;
 }
